backend/django_app/requirements.txt
# Django Dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-storages==1.14.2

# Database
psycopg2-binary==2.9.9

# AWS/S3
boto3==1.34.0

# Security
cryptography==41.0.8

# HTTP client
requests==2.31.0

# Production server
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0

# Utilities
Pillow==10.1.0

# FastAPI Dependencies (for AI microservice)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Dependencies
openai==1.3.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# File handling
aiofiles==23.2.1

# Additional utilities
pydantic==2.5.0
